advanced-data/streamlit-dashboarding/requirements.txt
python-dotenv
pandas
streamlit
altair
plotly.express
boto3
awswrangler